Update relative path of root index for generated api docs
There is a need for the new interface for these docs to
have nothing in the root directory. As a result, this commit
updates the location of the generated index.adoc file
so that it is instead generated in the rest_api/overview
directory. Additionally, this commit also updates the
relative path of the content links generated to point to
the correct directory for the index file now that it is nested.

Finally, this commit also resolves another issue with these
relative links: it replaces the periods in the generated
names based on the api schema and replaces them with
hyphens so that they can resolve properly.

Signed-off-by: kevinrizza <krizza@redhat.com>